server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
opencv-python==4.9.0.80
